requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.6
oauthlib==2.0.7
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
